Cast cast cast your bits gently for GCC. Closes #3652.

+++ machine/vm.hpp
@@ -232,7 +232,7 @@ namespace rubinius {
 
       if(stack_used < 0) stack_used = -stack_used;
 
-      if(stack_used > stack_size_) {
+      if(static_cast<size_t>(stack_used) > stack_size_) {
         raise_stack_error(state);
         return false;
       }
